# Use generator for single-threaded task generation