--- a/rio_tiler/tasks.py
+++ b/rio_tiler/tasks.py
@@ -8,7 +8,7 @@ from .constants import MAX_THREADS
 from .logger import logger
 from .models import ImageData
 
-TaskType = Sequence[Tuple[Union[futures.Future, Callable], str]]
+TaskType = Generator[Tuple[Union[futures.Future, Callable], str], None, None]
 
 
 def filter_tasks(
@@ -48,13 +48,15 @@ def create_tasks(reader: Callable, assets, threads, *args, **kwargs) -> TaskType
     if threads and threads > 1:
         logger.debug(f"Running tasks in ThreadPool with max_workers={threads}")
         with futures.ThreadPoolExecutor(max_workers=threads) as executor:
-            return [
+            yield from (
                 (executor.submit(reader, asset, *args, **kwargs), asset)
                 for asset in assets
-            ]
+            )
     else:
         logger.debug(f"Running tasks outside ThreadsPool (max_workers={threads})")
-        return [(partial(reader, asset, *args, **kwargs), asset) for asset in assets]
+        yield from (
+            (partial(reader, asset, *args, **kwargs), asset) for asset in assets
+        )
 
 
 def multi_arrays(