Edit: This PR should be a performance improvement for the `tasks` generator when using a single thread and for situations like a mosaic using `first` pixel selection, where you don't necessarily need to wait for _all_ the data returned from the generator.

-----

Original argument, since disproven with multithreaded code:

The issue is that here
https://github.com/cogeotiff/rio-tiler/blob/ff068185b269b4e556b4f922f0db6bca439763b9/rio_tiler/tasks.py#L50-L54

Since we use a list comprehension, the entire iterable will be "spent" and turned into a list before the first item will be yielded. Here's a contrived example with `sleep` to show my point:

```py
from concurrent import futures
from time import sleep

def create_tasks_map(threads=1):
    with futures.ThreadPoolExecutor(max_workers=threads) as executor:
        yield from executor.map(sleep, range(1, 4))

def create_tasks_gen(threads=1):
    with futures.ThreadPoolExecutor(max_workers=threads) as executor:
        yield from (
            executor.submit(sleep, i)
            for i in range(1, 4)
        )

def create_tasks_list(threads=1):
    with futures.ThreadPoolExecutor(max_workers=threads) as executor:
        return iter([
            executor.submit(sleep, i)
            for i in range(1, 4)
        ])

%time next(create_tasks_map())
# CPU times: user 851 µs, sys: 1.04 ms, total: 1.89 ms
# Wall time: 3 s

%time next(create_tasks_gen())
# CPU times: user 661 µs, sys: 927 µs, total: 1.59 ms
# Wall time: 1 s

%time next(create_tasks_list())
# CPU times: user 1.18 ms, sys: 1.11 ms, total: 2.29 ms
# Wall time: 6.01 s
```

Here you can see that when only loading the first item, the `yield from` is much faster because the list comprehension runs all the `sleep` calls before returning anything.

Side note, [it looks like under the hood `yield from` is pretty cool...](https://stackoverflow.com/a/26109157) 

> `yield from` ... establishes a transparent bidirectional connection between the caller and the sub-generator