@vincentsarago 2020-12-07
@kylebarron any update from the discussion/test over the week end ?