@kylebarron 2020-12-10
I think the conclusion from talking with @geospatial-jeff is that this is bad practice for the multithreaded case because it doesn't _send_ each item to the threadpool until the next item from the generator is called. So it's implicitly single-threaded.

But it would still be good to change the non-threaded code from a list to a generator.