@vincentsarago 2020-12-13
@kylebarron I'm still not sure on this one. Because we use partial to trigger the call to the `reader`, I don't see how using a generator will make the code run differently 

https://github.com/cogeotiff/rio-tiler/blob/5d8dc218e4a9dc2dbbeba514f234a304fd43eacc/rio_tiler/tasks.py#L40