@kylebarron 2020-12-13
😭 I really thought I was finding an improvement!

I didn't realize that `partial` doesn't _execute_ the function, it just wraps it, so the function isn't actually called until that `yield future(), asset` line